Factor out `UnsplashSearchContent`
