+++ desktop_photo_search/material/lib/main.dart
@@ -4,20 +4,16 @@
 
 import 'dart:io';
 
-import 'package:file_selector/file_selector.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_simple_treeview/flutter_simple_treeview.dart';
 import 'package:logging/logging.dart';
 import 'package:menubar/menubar.dart' as menubar;
 import 'package:provider/provider.dart';
 
 import 'src/model/photo_search_model.dart';
-import 'src/unsplash/photo.dart';
 import 'src/unsplash/unsplash.dart';
-import 'src/widgets/photo_details.dart';
 import 'src/widgets/photo_search_dialog.dart';
 import 'src/widgets/policy_dialog.dart';
-import 'src/widgets/split.dart';
+import 'src/widgets/unsplash_search_content.dart';
 import 'unsplash_access_key.dart';
 
 void main() {
@@ -58,16 +54,9 @@ class UnsplashSearchApp extends StatelessWidget {
   }
 }
 
-class UnsplashHomePage extends StatefulWidget {
+class UnsplashHomePage extends StatelessWidget {
   const UnsplashHomePage({required this.title, Key? key}) : super(key: key);
   final String title;
-
-  @override
-  State<UnsplashHomePage> createState() => _UnsplashHomePageState();
-}
-
-class _UnsplashHomePageState extends State<UnsplashHomePage> {
-  final _treeViewScrollController = ScrollController();
 
   @override
   Widget build(BuildContext context) {
@@ -100,51 +89,10 @@ class _UnsplashHomePageState extends State<UnsplashHomePage> {
 
     return Scaffold(
       appBar: AppBar(
-        title: Text(widget.title),
+        title: Text(title),
       ),
       body: photoSearchModel.entries.isNotEmpty
-          ? Split(
-              axis: Axis.horizontal,
-              initialFirstFraction: 0.4,
-              firstChild: Scrollbar(
-                controller: _treeViewScrollController,
-                child: SingleChildScrollView(
-                  controller: _treeViewScrollController,
-                  child: TreeView(
-                    nodes: photoSearchModel.entries
-                        .map(_buildSearchEntry)
-                        .toList(),
-                    indent: 0,
-                  ),
-                ),
-              ),
-              secondChild: Center(
-                child: photoSearchModel.selectedPhoto != null
-                    ? PhotoDetails(
-                        photo: photoSearchModel.selectedPhoto!,
-                        onPhotoSave: (photo) async {
-                          final path = await getSavePath(
-                            suggestedName: '${photo.id}.jpg',
-                            acceptedTypeGroups: [
-                              XTypeGroup(
-                                label: 'JPG',
-                                extensions: ['jpg'],
-                                mimeTypes: ['image/jpeg'],
-                              ),
-                            ],
-                          );
-                          if (path != null) {
-                            final fileData =
-                                await photoSearchModel.download(photo: photo);
-                            final photoFile = XFile.fromData(fileData,
-                                mimeType: 'image/jpeg');
-                            await photoFile.saveTo(path);
-                          }
-                        },
-                      )
-                    : Container(),
-              ),
-            )
+          ? const UnsplashSearchContent()
           : const Center(
               child: Text('Search for Photos using the Fab button'),
             ),
@@ -157,41 +105,6 @@ class _UnsplashHomePageState extends State<UnsplashHomePage> {
         tooltip: 'Search for a photo',
         child: const Icon(Icons.search),
       ),
-    );
-  }
-
-  TreeNode _buildSearchEntry(SearchEntry searchEntry) {
-    void selectPhoto(Photo photo) {
-      searchEntry.model.selectedPhoto = photo;
-    }
-
-    String labelForPhoto(Photo photo) => 'Photo by ${photo.user!.name}';
-
-    return TreeNode(
-      content: Expanded(
-        child: Text(searchEntry.query),
-      ),
-      children: searchEntry.photos
-          .map<TreeNode>(
-            (photo) => TreeNode(
-              content: Expanded(
-                child: Semantics(
-                  button: true,
-                  onTap: () => selectPhoto(photo),
-                  label: labelForPhoto(photo),
-                  excludeSemantics: true,
-                  child: InkWell(
-                    onTap: () => selectPhoto(photo),
-                    child: Padding(
-                      padding: const EdgeInsets.all(12.0),
-                      child: Text(labelForPhoto(photo)),
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          )
-          .toList(),
     );
   }
 }
